+++ clickgen/util.py
@@ -6,14 +6,12 @@
 .. moduleauthor:: Kaiz Khatri <kaizmandhu@gmail.com>
 """
 
-import functools
 import os
 import re
 import shutil
-import time
 from contextlib import contextmanager
 from pathlib import Path
-from typing import List, Set, Union
+from typing import List, Union
 
 from clickgen.db import DATA, CursorDB
 
@@ -23,8 +21,7 @@ def chdir(directory: Union[str, Path]):
     """Temporary change working directory using `with` syntax.
 
     :param directory: path to directory.
-    :type directory: Union[str, pathlib.Path]
-
+    :type directory: ``str`` or ``pathlib.Path``
     """
 
     prev_cwd = os.getcwd()
@@ -39,11 +36,12 @@ def remove_util(p: Union[str, Path]) -> None:
     """Remove this file, directory or symlink.
 
     :param p: path to directory.
-    :type p: Union[str, pathlib.Path]
-    :return: None
-    :rtype: None
+    :type p: ``str`` or ``pathlib.Path``
 
+    :return: None.
+    :rtype: ``None``
     """
+
     p_obj: Path = Path(p)
 
     if p_obj.exists():
@@ -65,13 +63,16 @@ class PNGProvider:
         """Init `PNGProvider`.
 
         :param bitmaps_dir: Path to directory where `.png` files are stored.
-        :type bitmaps_dir: Union[str, Path]
+        :type bitmaps_dir: ``str`` or ``pathlib.Path``
 
+        :returns: None.
+        :return: ``None``
         """
+
         super().__init__()
         self.bitmaps_dir = Path(bitmaps_dir)
-        for f in sorted(self.bitmaps_dir.iterdir()):
-            self.__pngs.append(f.name)
+        for file in sorted(self.bitmaps_dir.iterdir()):
+            self.__pngs.append(file.name)
 
         if len(self.__pngs) == 0:
             raise FileNotFoundError(
@@ -83,14 +84,16 @@ class PNGProvider:
 
         This method return file location in `pathlib.Path` object.
 
-        Runtime directory `sync` is **not supported**, Which means creating
-        or deleting a file on programs execution is not update class `__pngs`
+        Runtime directory `sync` is **not supported**, Which means creating \
+        or deleting a file on programs execution is not update class `__pngs` \
         state.
 
-        :param key: `key` is filename
-        :type key: str
-        :return: Returns `pathlib.Path` object or `list` of `pathlib.Path`
+        :param key: `key` is filename.
+        :type key: ``str``
+
+        :returns: Returns `pathlib.Path` object or `List` of `pathlib.Path`
                  object/s.
+        :rtype: ``List[pathlib.Path]`` or ``pathlib.Path``
         """
 
         k = key.split(".")
@@ -109,15 +112,34 @@ class PNGProvider:
 
 def add_missing_xcursors(
     directory: Path,
-    data: List[Set[str]] = DATA,
+    data: List = None,
     rename: bool = False,
     force: bool = False,
 ) -> None:
-    """Create symlinks of missing ``Xcursor``"""
+    """Create symlinks of missing ``Xcursor``.
+
+    :param directory: path to Xcursor directory.
+    :type directory: ``pathlib.Path``
+
+    :param data: Provide custom dataset of cursors. default to \
+    ~``clickgen.db.DATA``
+    :type data: ``List[set[str]]``
+
+    :param rename: If you want to rename ``Xcursor`` according to \
+        ``data``.
+    :type rename: ``bool``
+
+    :param force: To enable force renaming.
+    :type force: ``bool``
+    """
+
+    if data is None:
+        data = DATA
+
     if not directory.exists() or not directory.is_dir():
         raise NotADirectoryError(directory.absolute())
 
-    db: CursorDB = CursorDB(data)
+    cur_db: CursorDB = CursorDB(data)
 
     # Removing all symlinks cursors
     if force:
@@ -130,52 +152,13 @@ def add_missing_xcursors(
     for xcursor in xcursors:
         # Rename Xcursor according to Database, If necessary
         if rename:
-            new_path = db.rename_file(xcursor)
+            new_path = cur_db.rename_file(xcursor)
             if new_path:
                 xcursor = xcursor.rename(new_path)
 
         # Creating symlinks
-        links = db.search_symlinks(xcursor.stem)
+        links = cur_db.search_symlinks(xcursor.stem)
         if links:
             for link in links:
                 with chdir(directory):
                     os.symlink(xcursor, link)
-
-
-#
-# Useful for development
-#
-def timer(func):
-    """Print the runtime of the decorated function"""
-
-    @functools.wraps(func)
-    def wrapper_timer(*args, **kwargs):
-        start_time = time.perf_counter()
-
-        value = func(*args, **kwargs)
-
-        end_time = time.perf_counter()
-        run_time = end_time - start_time
-        print(f"Finished {func.__name__!r} in {run_time:.6f} secs")
-
-        return value
-
-    return wrapper_timer
-
-
-def debug(func):
-    """Print the function signature and return value"""
-
-    @functools.wraps(func)
-    def wrapper_debug(*args, **kwargs):
-        args_repr = [repr(a) for a in args]  # 1
-        kwargs_repr = [f"{k}={v!r}" for k, v in kwargs.items()]  # 2
-        signature = ", ".join(args_repr + kwargs_repr)  # 3
-
-        print(f"Calling {func.__name__}({signature})")
-        value = func(*args, **kwargs)
-        print(f"{func.__name__!r} returned {value!r}")  # 4
-
-        return value
-
-    return wrapper_debug